Improved handling of parsing GRPC arguments, and formatted
The significant change is to make the Parsed*Arguments structs for GRPC
arguments moveable, so that they can be constructed inside a try-catch
block and then moved outside of it. This allows the FunctionTier methods
to limit the scope of their try blocks to just the parse_grpc_args
functions, rather than enclosing everything in a try-catch. All other
files are only changed to improve formatting and readability.